pyproject.toml
[tool.poetry]
name = "rag-personal-assistant"
version = "0.1.0"
description = ""
authors = ["madDuarte <45204585+madDuarte@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
google-generativeai = "^0.8.5"
chromadb = "^1.3.5"
python-dotenv = "^1.2.1"
openai = "^2.8.1"
huggingface-hub = ">=0.34.0,<1.0"
numpy = "<2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
